changes based on Felipe's feedback
